% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agn.R
\docType{package}
\name{agn}
\alias{agn}
\alias{agn-package}
\title{\pkg{agn}}
\description{
\pkg{agn}
}
\author{
Australian Antarctic Data Centre \email{aadc@aad.gov.au}
}
\references{
\url{http://data.aad.gov.au/aadc/gaz/scar}
}

