% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{agn_near}
\alias{agn_near}
\title{Find placenames near a given location}
\usage{
agn_near(gaz, loc, max_distance)
}
\arguments{
\item{gaz}{data.frame: as returned by \code{\link{agn_read}}}

\item{loc}{numeric: longitude and latitude of target location}

\item{max_distance}{numeric: maximum search distance in kilometres}
}
\value{
data.frame of results
}
\description{
Find placenames near a given location
}
\examples{
\dontrun{
 g <- agn_read(cache_directory="c:/temp/gaz")
 agn_near(g,c(110,-66),10)

 ## using dplyr or magrittr
 g \%>\% agn_near(c(100,-66),20)
}

}
\references{
\url{http://www.scar.org/data-products/cga}
}
\seealso{
\code{\link{agn_read}}
}

