% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{agn_preferred}
\alias{agn_preferred}
\title{Find one name per feature according to specified preferences}
\usage{
agn_preferred(gaz, origin_country)
}
\arguments{
\item{gaz}{data.frame: as returned by \code{\link{agn_read}}}

\item{origin_country}{character: vector of country names. If a given feature has been named by one of these countries, this name will be returned. For valid values see \code{\link{agn_countries}}}
}
\value{
data.frame of results
}
\description{
Find one name per feature according to specified preferences
}
\examples{
\dontrun{
 g <- agn_read(cache_directory="c:/temp/gaz")
 g \%>\% agn_filter("^Ufs") \%>\% agn_preferred(origin_country=c("United States of America","Australia"))
}

}
\references{
\url{http://www.scar.org/data-products/cga}
}
\seealso{
\code{\link{agn_read}}
}

