% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{an_cache_directory}
\alias{an_cache_directory}
\title{The cache directory used by antanym}
\usage{
an_cache_directory(cache)
}
\arguments{
\item{cache}{string: the gazetteer data can be cached locally, so that it can be used offline later. Valid values are \code{"session"}, \code{"persistent"}, or a directory name. Specifying \code{cache="session"} will use a temporary directory that persists only for the current session. \code{cache="persistent"} will use \code{rappdirs::user_cache_dir()} to determine the appropriate directory to use. Otherwise, the input string will be assumed to be the path to the directory to use}
}
\value{
directory path
}
\description{
The cache directory used by antanym
}
\examples{
## per-session caching
an_cache_directory(cache = "session")

## persistent caching that will keep the data from one R session to the next
an_cache_directory(cache = "persistent")

}
\seealso{
\code{\link{an_read}}
}
