% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{an_feature_types}
\alias{an_feature_types}
\title{List feature types present in gazetteer data}
\usage{
an_feature_types(gaz)
}
\arguments{
\item{gaz}{data.frame or SpatialPointsDataFrame: as returned by \code{\link{an_read}}, \code{\link{an_preferred}}, or \code{\link{an_filter}}}
}
\value{
character vector of country names
}
\description{
The gazetteer place names are associated with different feature types (e.g. "Hill", "Mountain", "Water body"). This function lists the feature types that are present in a given data frame.
}
\examples{
\dontrun{
 g <- an_read(cache = "session")
 an_feature_types(g)
}
}
\seealso{
\code{\link{an_filter}} for filtering data according to feature type
}
