% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{an_preferred}
\alias{an_preferred}
\title{Find one name per feature according to specified preferences}
\usage{
an_preferred(gaz, origin_country)
}
\arguments{
\item{gaz}{data.frame: as returned by \code{\link{an_read}}}

\item{origin_country}{character: vector of country names. If a given feature has been named by one of these countries, this name will be returned. For valid values see \code{\link{an_countries}}}
}
\value{
data.frame of results
}
\description{
Find one name per feature according to specified preferences
}
\examples{
\dontrun{
 g <- an_read(cache_directory="c:/temp/gaz")

 ## get a single name per feature, preferring the
 ##  Polish name where there is one
 pnames <- an_preferred(g,"Poland")

 ## names starting with "Sm", preferring US names then
 ##  Australian ones (then whatever is available after that)
 g \%>\% an_filter("^Sm") \%>\% an_preferred(origin_country=c("United States of America","Australia"))
}

}
\references{
\url{http://www.scar.org/data-products/cga}
}
\seealso{
\code{\link{an_read}}
}

