% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antanym.R
\docType{package}
\name{antanym}
\alias{antanym}
\alias{antanym-package}
\title{\pkg{antanym}}
\description{
\pkg{antanym}
}
\author{
Australian Antarctic Data Centre \email{aadc@aad.gov.au}
}
\references{
\url{http://data.aad.gov.au/aadc/gaz/scar}
}

