% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cga.R
\docType{package}
\name{cga}
\alias{cga}
\alias{cga-package}
\title{\pkg{cga}}
\description{
\pkg{cga}
}
\author{
Australian Antarctic Data Centre \email{aadc@aad.gov.au}
}
\references{
\url{http://data.aad.gov.au/aadc/gaz/scar}
}

