% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{names_near}
\alias{names_near}
\title{Find placenames near a given location}
\usage{
names_near(cga, loc, max_distance)
}
\arguments{
\item{cga}{SQLiteConnection: as returned by \code{\link{load_cga}}}

\item{loc}{numeric: longitude and latitude of target location}

\item{max_distance}{numeric: maximum search distance in kilometres}
}
\value{
data.frame of results
}
\description{
Find placenames near a given location
}
\examples{
\dontrun{
 g <- load_cga(cache_directory="c:/temp/cga")
 names_near(g,c(110,-66),10)

 ## using dplyr or magrittr
 gg \%>\% names_near(c(100,-66),10)
}

}
\references{
\url{http://www.scar.org/data-products/cga}
}
\seealso{
\code{\link{load_cga}}
}

