% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_names}
\alias{cga_countries}
\alias{cga_gazetteers}
\alias{search_names}
\title{Search for place names}
\usage{
search_names(cga, query, feature_type, origin_country, origin_gazetteer)

cga_countries(cga)

cga_gazetteers(cga)
}
\arguments{
\item{cga}{SQLiteDriver: as returned by \code{\link{load_cga}}}

\item{query}{string: SQL-syntax search string. Matches are case-insensitive}

\item{feature_type}{string: if provided, search only for place names corresponding to features of this type}

\item{origin_country}{string: if provided, search only for place names originating from this country. For valid entries see \code{\link{cga_countries}}}

\item{origin_gazetteer}{string: if provided, search only for place names originating from this source gazetteer. For valid entries see \code{\link{cga_gazetteers}}}
}
\value{
data.frame of results
}
\description{
Search for place names
}
\examples{
\dontrun{
 g <- load_cga(cache_directory="c:/temp/cga")
 search_names(g,"Ufs\%")
 search_names(g,"Ufs\%",feature_type="Island")
 search_names(g,"Ufs\%",feature_type="Island",origin_country="Australia")
}
}
\references{
\url{http://www.scar.org/data-products/cga}
}
\seealso{
\code{\link{load_cga}}
}

