% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{agn_filter}
\alias{agn_filter}
\alias{cga_countries}
\alias{cga_display_scales}
\alias{cga_feature_types}
\alias{cga_gazetteers}
\title{Subset place names by various criteria}
\usage{
agn_filter(cga, query, extent, feature_type, origin_country, origin_gazetteer,
  display_scale)

cga_countries(cga)

cga_feature_types(cga)

cga_gazetteers(cga)

cga_display_scales(cga)
}
\arguments{
\item{cga}{data.frame: as returned by \code{\link{agn_read}}}

\item{query}{string: regular expression to match on place name. Matches are case-insensitive}

\item{extent}{raster::extent object or vector of c(xmin,xmax,ymin,ymax): if provided, search only for names within this bounding box}

\item{feature_type}{string: if provided, search only for place names corresponding to features of this type. For valid values see \code{\link{cga_feature_types}}}

\item{origin_country}{string: if provided, search only for place names originating from this country. For valid values see \code{\link{cga_countries}}}

\item{origin_gazetteer}{string: if provided, search only for place names originating from this source gazetteer. For valid values see \code{\link{cga_gazetteers}}}

\item{display_scale}{string: if provided, search only for place names that have been marked for display at this map scale. For valid values see \code{\link{cga_display_scales}}}
}
\value{
data.frame of results
}
\description{
Subset place names by various criteria
}
\examples{
\dontrun{
 g <- agn_read(cache_directory="c:/temp/cga")
 agn_filter(g,"Ufs")
 agn_filter(g,"Ufs",feature_type="Island")
 agn_filter(g,"Ufs",feature_type="Island",origin_country="Australia")

 nms <- agn_filter(g,extent=c(100,120,-70,-65),display_scale="2000000",
    origin_country="Australia")
 with(nms,plot(longitude,latitude))
 with(nms,text(longitude,latitude,place_name))

 ## using dplyr or magrittr
 g \%>\% agn_filter("Ross",feature_type="Ice shelf")
 g \%>\% agn_near(c(100,-66),20) \%>\% agn_filter(feature_type="Island")
}
}
\references{
\url{http://www.scar.org/data-products/cga}
}
\seealso{
\code{\link{agn_read}}
}

