% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{an_filter}
\alias{an_cga_sources}
\alias{an_countries}
\alias{an_feature_types}
\alias{an_filter}
\title{Subset place names by various criteria}
\usage{
an_filter(gaz, query, extent, feature_type, origin_country, origin_gazetteer,
  cga_source)

an_countries(gaz)

an_feature_types(gaz)

an_cga_sources(gaz)
}
\arguments{
\item{gaz}{data.frame or SpatialPointsDataFrame: as returned by \code{\link{an_read}}}

\item{query}{string, regular expression: Return only place names matching this pattern}

\item{extent}{raster Extent object or vector of c(longitude_min,longitude_max,latitude_min,latitude_max): if provided, search only for names within this bounding box}

\item{feature_type}{string, regular expression: Return only place names corresponding to feature types matching this pattern. For valid feature type names see \code{\link{an_feature_types}}}

\item{origin_country}{string, regular expression: Return only names originating from this countries matching this pattern. For valid country names see \code{\link{an_countries}}}

\item{origin_gazetteer}{string, regular expression: Return only place names originating from gazetteers matching this pattern. For valid gazetteer names see \code{\link{an_gazetteers}}}

\item{cga_source}{string, regular expression: Return only place names in the SCAR Composite Gazetteer originating from contributing gazetteers matching this pattern. For valid CGA source gazetteer names see \code{\link{an_cga_sources}}. (Only applicable to the SCAR Composite Gazetteer)}
}
\value{
data.frame of results
}
\description{
All matches are case-insensitive.
}
\examples{
\dontrun{
 g <- an_read(cache_directory="c:/temp/gaz")
 an_filter(g,"Ufs")
 an_filter(g,"Ufs",feature_type="Island")
 an_filter(g,"Ufs",feature_type="Island",origin_country="Australia|United States of America")

 nms <- an_filter(g,extent=c(100,120,-70,-65),
    origin_country="Australia")
 with(nms,plot(longitude,latitude))
 with(nms,text(longitude,latitude,place_name))

 ## using dplyr or magrittr
 g \%>\% an_filter("Ross",feature_type="Ice shelf|Mountain")
 g \%>\% an_near(c(100,-66),20) \%>\% an_filter(feature_type="Island")
}
}
\references{
\url{http://www.scar.org/data-products/cga}
}
\seealso{
\code{\link{an_read}}
}

