% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gazetteers.R
\name{an_gazetteers}
\alias{an_gazetteers}
\title{The place name gazetteers available}
\usage{
an_gazetteers(gaz)
}
\arguments{
\item{gaz}{data.frame or SpatialPointsDataFrame: (optional) as returned by \code{\link{an_read}}, \code{\link{an_preferred}}, or \code{\link{an_filter}}}
}
\value{
character vector. If \code{gaz} was provided, this will be a list of all gazetteers present in \code{gaz}. Otherwise, it will be a list of all gazetteers available through the antanym package
}
\description{
Return a character vector that lists all of the gazetteers present in the \code{gaz} data, or (if \code{gaz} was not provided) all of the gazetteers available through the antanym package. Currently only one gazetteer is available: the Composite Gazetteer of Antarctica.
}
\examples{

an_gazetteers()

}
\seealso{
\code{\link{an_read}}, \code{\link{an_filter}}
}
