% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{an_mapscale}
\alias{an_mapscale}
\title{Calculate approximate map scale}
\usage{
an_mapscale(map_dimensions, map_extent)
}
\arguments{
\item{map_dimensions}{numeric: 2-element numeric giving width and height of the map, in mm}

\item{map_extent}{vector of c(longitude_min, longitude_max, latitude_min, latitude_max): the geographic extent of the map. \code{map_extent} can also be passed as a raster Extent object, a Raster object (in which case its extent will be used), a Spatial object (in which case the bounding box of the object will be used as the extent), or a matrix (in which case it will be assumed to be the output of \code{sp::bbox})}
}
\value{
numeric
}
\description{
Calculate approximate map scale
}
\examples{
## an A3-sized map of the Southern Ocean (1:20M)
an_mapscale(map_dimensions = c(400, 570), map_extent = c(-180, 180, -90, -40))

}
