% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{an_near}
\alias{an_near}
\title{Find placenames near a given location}
\usage{
an_near(gaz, loc, max_distance)
}
\arguments{
\item{gaz}{data.frame or SpatialPointsDataFrame: as returned by \code{\link{an_read}}, \code{\link{an_preferred}}, or \code{\link{an_filter}}}

\item{loc}{numeric: target location (a two-element numeric vector giving longitude and latitude, or a SpatialPoints object)}

\item{max_distance}{numeric: maximum search distance in kilometres}
}
\value{
data.frame of results
}
\description{
Find placenames near a given location
}
\examples{
\dontrun{
 g <- an_read(cache = "session")

 ## named features within 10km of 110E, 66S
 an_near(g, c(110, -66), 10)

 ## using pipe operator
 g \%>\% an_near(c(100, -66), 10)

 ## with sp objects
 gsp <- an_read(cache = "session", sp = TRUE)
 loc <- SpatialPoints(matrix(c(110, -66), nrow = 1),
   proj4string=CRS("+proj=longlat +datum=WGS84 +ellps=WGS84"))
 an_near(gsp, loc, 10)
}

}
\references{
\url{https://www.scar.org/data-products/place-names/}
}
\seealso{
\code{\link{an_read}}
}
