% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cga_preferred_names.R
\name{an_preferred}
\alias{an_preferred}
\title{Find one name per feature in the SCAR Composite Gazetteer}
\usage{
an_preferred(gaz, origin_country, cga_source, unmatched = "random")
}
\arguments{
\item{gaz}{data.frame or SpatialPointsDataFrame: as returned by \code{\link{an_read}}}

\item{origin_country}{character: deprecated. Use \code{cga_source} instead}

\item{cga_source}{character: vector of source gazetteers, in order of preference. If a given feature has been named in one of these gazetteers, this place name will be chosen. If the feature in question has not been given a name in any of these gazetteers, a place name given in another gazetteer will be chosen, with preference according to the \code{unmatched} parameter. For valid \code{cga_source} values, see \code{\link{an_cga_sources}}}

\item{unmatched}{string: how should names be chosen for features that do not have a name in one of the preferred \code{cga_source} gazetteers? Valid values are "random" (the non-preferred gazetteers will be randomly ordered) or "count" (the non-preferred gazetteers will be ordered by their number of entries, with the largest gazetteer first)}
}
\value{
data.frame of results
}
\description{
The SCAR Composite Gazetteer of Antarctic is a compilation of place names provided by many different countries. The composite nature of the CGA means that there may be multiple names associated with a single feature. The \code{an_preferred} function can be used to resolve a single name per feature. Specify one or more \code{cga_source} entries and the input \code{gaz} will be filtered to a single name per feature. For features that have multiple names (i.e. appear in multiple source gazetteers), a single name will be chosen, preferring names from the specified \code{cga_source} entries where possible.
}
\examples{
\dontrun{
 g <- an_read(cache = "session")

 ## get a single name per feature, preferring the
 ##  Polish name where there is one
 pnames <- an_preferred(g, cga_source = "POL")

 ## names starting with "Sm", preferring US names then
 ##  Australian ones if available
 g \%>\% an_filter("^Sm") \%>\% an_preferred(cga_source = c("USA", "AUS"))
}

}
\references{
\url{http://www.scar.org/data-products/cga}
}
\seealso{
\code{\link{an_read}}, \code{\link{an_cga_sources}}
}
