% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{an_read}
\alias{an_read}
\alias{an_gazetteers}
\title{Load Antarctic and subantarctic place name gazetteers}
\usage{
an_read(gazetteers = "all", sp = FALSE, cache_directory,
  refresh_cache = FALSE, verbose = FALSE)

an_gazetteers()
}
\arguments{
\item{gazetteers}{character: vector of gazetteers to load. For the list of available gazetteers, see \code{\link{an_gazetteers}}. Use \code{gazetteers="all"} to load all available gazetteers}

\item{sp}{logical: if FALSE return a data.frame; if TRUE return a SpatialPointsDataFrame}

\item{cache_directory}{string: (optional) cache the gazetteer data file locally in this directory, so that it can be used offline later. The cache directory will be created if it does not exist. A warning will be given if a cached copy exists and is more than 30 days old}

\item{refresh_cache}{logical: if TRUE, and a data file already exists in the cache_directory, it will be refreshed. If FALSE, the cached copy will be used}

\item{verbose}{logical: show progress messages?}
}
\value{
a data.frame or SpatialPointsDataFrame
}
\description{
Load Antarctic and subantarctic place name gazetteers
}
\examples{
\dontrun{
 g <- an_read(cache_directory="c:/temp/gaz")
}

}
\references{
\url{http://www.scar.org/data-products/cga} \url{http://data.aad.gov.au/aadc/gaz/}
}
