% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{an_suggest}
\alias{an_suggest}
\title{Suggest names for a map (experimental)}
\usage{
an_suggest(gaz, map_scale, map_extent, map_dimensions)
}
\arguments{
\item{gaz}{data.frame or SpatialPointsDataFrame: as returned by \code{\link{an_read}}}

\item{map_scale}{numeric: the scale of the map (e.g. 20e6 for a 1:20M map). If \code{map_scale} is not provided, it will be estimated from \code{map_extent} and \code{map_dimensions}}

\item{map_extent}{raster Extent object or vector of c(longitude_min, longitude_max, latitude_min, latitude_max): the extent of the area for which name suggestions are sought. Not required if \code{map_scale} is provided}

\item{map_dimensions}{numeric: 2-element numeric giving width and height of the map, in mm. Not required if \code{map_scale} is provided}
}
\value{
data.frame of names with a "score" column added. The data.frame will be sorted in descending score order. Names with higher scores are those that are suggested as the most suitable for display.
}
\description{
Features are given a suitability score based on how often (and at what map scales) they have been named on maps prepared by expert cartographers.
This is an experimental function and currently only implemented for \code{map_scale} values of 10 million or larger.
}
\examples{
\dontrun{
 g <- an_read(cache_directory="c:/temp/gaz")

 ## get a single name per feature, preferring the
 ##  Australian name where there is one
 g <- an_preferred(g, "Australia")

 ## suggested names for a 100x100 mm map covering 60-90E, 70-60S
 ##  (this is about a 1:12M scale map)

 suggested <- an_suggest(g, map_extent = c(60, 90, -70, -60),map_dimensions = c(100, 100))
 head(suggested, 20) ## top 20 names
}
}
\seealso{
\code{\link{an_read}} \code{\link{an_thin}}
}
