% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{an_thin}
\alias{an_thin}
\title{Thin names to give approximately uniform spatial coverage}
\usage{
an_thin(gaz, n, score_col = "score", score_weighting = 5)
}
\arguments{
\item{gaz}{data.frame: as returned by \code{\link{an_read}}}

\item{n}{numeric: number of names to return}

\item{score_col}{string: the name of the column that gives the relative score (e.g. as returned by \code{an_suggest})}

\item{score_weighting}{numeric: weighting of scores relative to spatial distribution. A lower \code{score_weighting} value will tend to choose lower-scored names
in order to achieve better spatial uniformity. A higher \code{score_weighting} value will trade spatial uniformity in favour of selecting
higher-scored names}
}
\value{
data.frame
}
\description{
Thin names to give approximately uniform spatial coverage
}
\examples{
\dontrun{
 g <- an_read(cache_directory="c:/temp/gaz")

 ## get a single name per feature, preferring the
 ##  Australian name where there is one
 g <- an_preferred(g,"Australia")

 ## suggested names for a 100x100 mm map covering 60-90E, 70-60S
 ##  (this is about a 1:12M scale map)
 suggested <- an_suggest(g,map_extent=c(60,90,-70,-60),map_dimensions=c(100,100))
 head(suggested,20) ## top 20 names by score
 an_thin(suggested,20) ## 20 names chosen for spatial coverage and score
}

}
\seealso{
\code{\link{an_read}} \code{\link{an_suggest}}
}

