% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antanym.R
\docType{package}
\name{antanym}
\alias{antanym}
\alias{antanym-package}
\alias{antanym-package}
\title{\pkg{antanym}}
\description{
Antarctic geographic place names from the SCAR Composite Gazetteer of Antarctica, and functions for working with those place names.
}
\references{
\url{http://data.aad.gov.au/aadc/gaz/scar}
}
