% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_cga}
\alias{load_cga}
\title{Load the Composite Gazetteer of Antarctica}
\usage{
load_cga(cache_directory, refresh_cache = FALSE, verbose = TRUE)
}
\arguments{
\item{cache_directory}{string: (optional) cache the CGA data file locally in this directory, so that it can be used offline later. The cache directory will be created if it does not exist}

\item{refresh_cache}{logical: if TRUE, and a data file already exists in the cache_directory, it will be refreshed. If FALSE, the cached copy will be used}

\item{verbose}{logical: show progress messages?}
}
\value{
An object of class \code{SQLiteConnection}
}
\description{
Load the Composite Gazetteer of Antarctica
}
\examples{
\dontrun{
 g <- load_cga(cache_directory="c:/temp/cga")
}

}
\references{
\url{http://www.scar.org/data-products/cga}
}

