% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_names}
\alias{cga_countries}
\alias{cga_display_scales}
\alias{cga_feature_types}
\alias{cga_gazetteers}
\alias{search_names}
\title{Search for place names}
\usage{
search_names(cga, query, feature_type, origin_country, origin_gazetteer,
  display_scale)

cga_countries(cga)

cga_feature_types(cga)

cga_gazetteers(cga)

cga_display_scales(cga)
}
\arguments{
\item{cga}{SQLiteDriver: as returned by \code{\link{load_cga}}}

\item{query}{string: SQL-syntax search string. Matches are case-insensitive}

\item{feature_type}{string: if provided, search only for place names corresponding to features of this type. For valid values see \code{\link{cga_feature_types}}}

\item{origin_country}{string: if provided, search only for place names originating from this country. For valid values see \code{\link{cga_countries}}}

\item{origin_gazetteer}{string: if provided, search only for place names originating from this source gazetteer. For valid values see \code{\link{cga_gazetteers}}}

\item{display_scale}{string: if provided, search only for place names that have been marked for display at this map scale. For valid values see \code{\link{cga_display_scales}}}
}
\value{
data.frame of results
}
\description{
Search for place names
}
\examples{
\dontrun{
 g <- load_cga(cache_directory="c:/temp/cga")
 search_names(g,"Ufs\%")
 search_names(g,"Ufs\%",feature_type="Island")
 search_names(g,"Ufs\%",feature_type="Island",origin_country="Australia")

 ## using dplyr or magrittr
 g \%>\% search_names("Ross\%",feature_type="Ice shelf")

 nms <- search_names(gg,display_scale="2000000",origin_country="Australia")
 with(nms,plot(longitude,latitude))
 with(nms,text(longitude,latitude,place_name))
}
}
\references{
\url{http://www.scar.org/data-products/cga}
}
\seealso{
\code{\link{load_cga}}
}

